Thread view component now requests a repaint for the stuff it hasn't managed to draw yet
This means you can now responsively move around the UI, even if you are viewing the entire damn dump of data!

Also fixed TimelineComponent to immediately erase the old cursor location - because we should be able to guarantee that we are on the EDT.

Can we also draw the new one immeditately?